# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{slug}/tickets": {
      "post": {
        "operationId": "TicketsController_create",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new ticket",
        "tags": [
          "tickets"
        ]
      },
      "get": {
        "operationId": "TicketsController_findAll",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "current",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Page size",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CREATED",
                "VERIFIED",
                "IN_PROGRESS",
                "VERIFY_FIX",
                "CLOSED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BUG",
                "ENHANCEMENT",
                "TASK",
                "QUESTION"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            }
          },
          {
            "name": "assignedTo",
            "required": false,
            "in": "query",
            "description": "User id of the assignee, or \"self\" for the caller (user or agent)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unassigned",
            "required": false,
            "in": "query",
            "description": "Only tickets with no user or agent assignee",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page of tickets: { total, current, size, hasNext, hasPrev, records }"
          },
          "400": {
            "description": "Invalid filter or paging parameter"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tickets for a project (paginated)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}": {
      "get": {
        "operationId": "TicketsController_findByRef",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a ticket by reference (KODA-42 or CUID)",
        "tags": [
          "tickets"
        ]
      },
      "patch": {
        "operationId": "TicketsController_update",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a ticket",
        "tags": [
          "tickets"
        ]
      },
      "delete": {
        "operationId": "TicketsController_softDelete",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - admin role required"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soft delete a ticket (admin only)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/assign": {
      "post": {
        "operationId": "TicketsController_assign",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot assign to both user and agent"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign or unassign a ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/verify": {
      "post": {
        "operationId": "TicketsController_verify",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify a ticket (CREATED → VERIFIED)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/start": {
      "post": {
        "operationId": "TicketsController_start",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Start work on a ticket (VERIFIED → IN_PROGRESS)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/fix": {
      "post": {
        "operationId": "TicketsController_fix",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fix submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit fix for verification (IN_PROGRESS → VERIFY_FIX)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/verify-fix": {
      "post": {
        "operationId": "TicketsController_verifyFix",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Fix reviewed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject fix (VERIFY_FIX → CLOSED or IN_PROGRESS)",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/close": {
      "post": {
        "operationId": "TicketsController_close",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ticket closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Close a ticket",
        "tags": [
          "tickets"
        ]
      }
    },
    "/api/projects/{slug}/tickets/{ref}/reject": {
      "post": {
        "operationId": "TicketsController_reject",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ref",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionWithCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transition"
          },
          "404": {
            "description": "Ticket or project not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a ticket (CREATED or VERIFIED → REJECTED)",
        "tags": [
          "tickets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Koda API",
    "description": "Dev ticket tracker API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "type": {
            "enum": [
              "BUG",
              "ENHANCEMENT",
              "TASK",
              "QUESTION"
            ],
            "type": "string",
            "description": "Ticket type",
            "example": "BUG"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "description": "Ticket title",
            "example": "Fix login bug"
          },
          "description": {
            "type": "string",
            "description": "Ticket description",
            "example": "Users cannot login with email"
          },
          "priority": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "type": "string",
            "description": "Ticket priority",
            "example": "MEDIUM",
            "default": "MEDIUM"
          }
        },
        "required": [
          "type",
          "title"
        ]
      },
...
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "description": "Ticket title",
            "example": "Updated ticket title"
          },
          "description": {
            "type": "string",
            "description": "Ticket description",
            "example": "Updated description"
          },
          "priority": {
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "type": "string",
            "description": "Ticket priority",
            "example": "HIGH"
          },
          "status": {
            "enum": [
              "CREATED",
              "VERIFIED",
              "IN_PROGRESS",
              "VERIFY_FIX",
              "CLOSED",
              "REJECTED"
            ],
            "type": "string",
            "description": "Ticket status",
            "example": "IN_PROGRESS"
          }
        }
      },
      "AssignTicketDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID (CUID) to assign the ticket to"
          },
          "agentId": {
            "type": "string",
            "description": "Agent ID (CUID) to assign the ticket to"
          }
        }
      },
      "TransitionWithCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "description": "Comment body explaining the transition",
            "example": "This ticket has been verified and is ready to work on"
          }
        }
      },
...
    }
  }
}